+++ server/recursiveShare_test.go
@@ -1,6 +1,8 @@
 package server_test
 
 import (
+	"bytes"
+	"io/ioutil"
 	"strings"
 	"testing"
 	"time"
@@ -10,34 +12,35 @@ import (
 	"decipher.com/object-drive-server/util/testhelpers"
 )
 
-func TestShareRecursive(t *testing.T) {
-
-	/* resource strings for sharing
-	fakeDN0 = `cn=test tester10,ou=people,ou=dae,ou=chimera,o=u.s. government,c=us`
-	fakeDN1 = `cn=test tester01,ou=people,ou=dae,ou=chimera,o=u.s. government,c=us`
-	fakeDN2 = `cn=test tester02,ou=people,ou=dae,ou=chimera,o=u.s. government,c=us`
-	newOwner := "user/" + fakeDN1
-	*/
+func TestUpdateObjectShareRecursive(t *testing.T) {
 
 	cases := []struct {
-		// creator and requester are indexes into global clients array
+		// Scenario:
+		// 1. creator creates an hierarchy of object with no permissions
+		//    and initialACM: root/child1Obj/child2Obj/child3Obj
+		// 2. sharer then updates child 1 with updatedACM and perms.
+		// 3. requester tries to get child 1, expecting success
+		// 4. requester tries to get child 3, expecting success
+
+		// identities are indexes into the global clients array
 		creator   int
 		sharer    int
 		requester int
 
-		// object properties
-		classif string
-		perms   protocol.Permission
-		// shareTo changes?
-		shareTo string
-		// TODO(cm) abstract ACM up here instead of hardcoding below.
+		// Updated permissions for UpdateObject. None are passed to CreateObject.
+		perms protocol.Permission
+
+		// initialACM is given to CreateObject
+		initialACM string
+
+		// updatedACM is given to UpdateObject.
+		updatedACM string
 	}{
 		{
 			0, // creator
 			0, // sharer
 			1, // requester
-			"U",
-			protocol.Permission{
+			protocol.Permission{ // updated permissions. None are passed at first.
 				Create: protocol.PermissionCapability{
 					AllowedResources: []string{
 						"user/" + fakeDN1 + "/test tester01",
@@ -64,14 +67,9 @@ func TestShareRecursive(t *testing.T) {
 					},
 				},
 			},
-			"",
+			testhelpers.ValidACMUnclassifiedFOUOSharedToTester10,      // initial acm
+			testhelpers.ValidACMUnclassifiedFOUOSharedToTester01And02, // updated acm
 		},
-	}
-
-	// quick func to make random names
-	randomName := func(name string) string {
-		s, _ := util.NewGUID()
-		return name + s
 	}
 
 	for _, c := range cases {
@@ -83,8 +81,7 @@ func TestShareRecursive(t *testing.T) {
 		cor := protocol.CreateObjectRequest{
 			NamePathDelimiter: ":::",
 			Name:              strings.Join([]string{root, child1, child2, child3}, ":::"),
-			// RawAcm:            testhelpers.ValidACMUnclassifiedFOUOSharedToTester01And02,
-			RawAcm: testhelpers.ValidACMUnclassifiedFOUOSharedToTester10,
+			RawAcm:            c.initialACM,
 		}
 		// child3 is returned
 		child3Obj, err := clients[c.creator].C.CreateObject(cor, nil)
@@ -102,13 +99,11 @@ func TestShareRecursive(t *testing.T) {
 		}
 
 		uor := protocol.UpdateObjectRequest{
-			ID:            child1Obj.ID,
-			ChangeToken:   child1Obj.ChangeToken,
+			ID:             child1Obj.ID,
+			ChangeToken:    child1Obj.ChangeToken,
 			RecursiveShare: true,
-			Permission:    c.perms,
-			// We HAVE to use an ACM? See issue 775
-			RawAcm: testhelpers.ValidACMUnclassifiedFOUOSharedToTester01And02,
-			// RawAcm: testhelpers.ValidACMUnclassifiedFOUOSharedToTester10,
+			Permission:     c.perms,
+			RawAcm:         c.updatedACM,
 		}
 		clients[c.sharer].C.Verbose = testing.Verbose()
 		_, err = clients[c.sharer].C.UpdateObject(uor)
@@ -141,7 +136,157 @@ func TestShareRecursive(t *testing.T) {
 			}
 			break
 		}
+	}
+}
 
+func TestUpdateObjectAndStreamShareRecursive(t *testing.T) {
+
+	cases := []struct {
+		// Scenario:
+		// 1. creator creates an hierarchy of object with no permissions
+		//    and initialACM: root/child1Obj/child2Obj/child3Obj
+		//    Note that child3Obj is created with a filestream.
+		// 2. sharer then updates child 1 with updatedACM and perms.
+		// 3. requester tries to get child 1's object stream, expecting success
+		// 4. requester tries to get child 3's object stream, expecting success
+
+		// identies are indexes into the global clients array
+		creator   int
+		sharer    int
+		requester int
+
+		// Updated permissions for UpdateObjectAndStream. None are passed to CreateObject.
+		perms protocol.Permission
+
+		// streamContents is copied into an io.Reader for UpdateObjectAndStream
+		streamContents []byte
+
+		// initialACM is given to UpdateObjectAndStream.
+		initialACM string
+
+		// updatedACM is given to UpdateObjectAndStream.
+		updatedACM string
+	}{
+		{
+			0, // creator
+			0, // sharer
+			1, // requester
+			protocol.Permission{ // updated permissions. None are passed at first.
+				Create: protocol.PermissionCapability{
+					AllowedResources: []string{
+						"user/" + fakeDN1 + "/test tester01",
+					},
+				},
+				Read: protocol.PermissionCapability{
+					AllowedResources: []string{
+						"user/" + fakeDN1 + "/test tester01",
+					},
+				},
+				Update: protocol.PermissionCapability{
+					AllowedResources: []string{
+						"user/" + fakeDN1 + "/test tester01",
+					},
+				},
+				Delete: protocol.PermissionCapability{
+					AllowedResources: []string{
+						"user/" + fakeDN1 + "/test tester01",
+					},
+				},
+				Share: protocol.PermissionCapability{
+					AllowedResources: []string{
+						"user/" + fakeDN1 + "/test tester01",
+					},
+				},
+			},
+			[]byte("The Biologist"),
+			testhelpers.ValidACMUnclassifiedFOUOSharedToTester10,      // initial acm
+			testhelpers.ValidACMUnclassifiedFOUOSharedToTester01And02, // updated acm
+		},
 	}
 
+	for _, c := range cases {
+		root, child1, child2, child3 := randomName("root"), randomName("child1"), randomName("child2"), randomName("child3")
+
+		t.Logf("Create object hierarchy:\n root: %s\n child1: %s\n child2: %s\n child3: %s\n",
+			root, child1, child2, child3)
+
+		cor := protocol.CreateObjectRequest{
+			NamePathDelimiter: ":::",
+			Name:              strings.Join([]string{root, child1, child2, child3}, ":::"),
+			RawAcm:            c.initialACM,
+		}
+		// child3 is returned
+		child3Obj, err := clients[c.creator].C.CreateObject(cor, bytes.NewBuffer(c.streamContents))
+		if err != nil {
+			t.Errorf("create failed for creator %v: %v", c.requester, err)
+		}
+		// get the parentid of the parentid
+		child2Obj, err := clients[c.creator].C.GetObject(child3Obj.ParentID)
+		if err != nil {
+			t.Errorf("creator could not get child2: %v", err)
+		}
+		child1Obj, err := clients[c.creator].C.GetObject(child2Obj.ParentID)
+		if err != nil {
+			t.Errorf("expected creator to have access to child1: %v", err)
+		}
+
+		uor := protocol.UpdateObjectAndStreamRequest{
+			ID: child1Obj.ID,
+			// NOTE(cm): If we do not supply a name here, we get
+			// "status":400,"message":"file must be supplied as multipart mime part"
+			// uploadDownload.go line 186
+			// Is this a bug where we are not preserving the original name?
+			Name:           "should_i_need_a_name",
+			ChangeToken:    child1Obj.ChangeToken,
+			RecursiveShare: true,
+			Permission:     c.perms,
+			RawAcm:         c.updatedACM,
+		}
+		clients[c.sharer].C.Verbose = testing.Verbose()
+		_, err = clients[c.sharer].C.UpdateObjectAndStream(uor, bytes.NewBuffer(c.streamContents))
+		if err != nil {
+			t.Errorf("error updating object: %v", err)
+			t.FailNow()
+		}
+
+		// can requester get child 1 stream?
+		r, err := clients[c.requester].C.GetObjectStream(child1Obj.ID)
+		if err != nil {
+			t.Errorf("requester could not get child1: %v", err)
+			t.FailNow()
+		}
+		contents, _ := ioutil.ReadAll(r)
+		if bytes.Compare(contents, c.streamContents) != 0 {
+			t.Errorf("expected stream contents to be: %s\nbut got: %s",
+				string(c.streamContents), string(contents))
+		}
+
+		// can requester get object child 3?
+		tries := 0
+		for {
+			// We must retry because this is an async operation on the server.
+			r, err := clients[c.requester].C.GetObjectStream(child3Obj.ID)
+			if err != nil {
+				if tries < 50 {
+					tries++
+					t.Logf("Sleeping 50 ms. Tries %v", tries)
+					time.Sleep(50 * time.Millisecond)
+					continue
+				}
+				t.Errorf("GetObject should succeed for tester1 on child3: %v", err)
+				t.FailNow()
+			}
+			contents, _ = ioutil.ReadAll(r)
+			if bytes.Compare(contents, c.streamContents) != 0 {
+				t.Errorf("expected stream contents to be: %s\nbut got: %s",
+					string(c.streamContents), string(contents))
+			}
+			break
+		}
+	}
+}
+
+func randomName(name string) string {
+	s, _ := util.NewGUID()
+	return name + s
 }